0.1

+++ resources/views/users/index.blade.php
@@ -6,7 +6,9 @@
             <h2>Users Management</h2>
         </div>
         <div class="pull-right">
+            @can('user-create')
             <a class="btn btn-success" href="{{ route('users.create') }}"> Create New User</a>
+            @endcan
         </div>
     </div>
 </div>
@@ -15,6 +17,7 @@
     <p>{{ $message }}</p>
 </div>
 @endif
+<br>
 <table class="table table-bordered">
     <tr>
         <th>No</th>
@@ -37,14 +40,17 @@
         </td>
         <td>
             <a class="btn btn-info" href="{{ route('users.show',$user->id) }}">Show</a>
+            @can('user-edit')
             <a class="btn btn-primary" href="{{ route('users.edit',$user->id) }}">Edit</a>
+            @endcan
+            @can('user-delete')
                 {!! Form::open(['method' => 'DELETE','route' => ['users.destroy', $user->id],'style'=>'display:inline']) !!}
                     {!! Form::submit('Delete', ['class' => 'btn btn-danger']) !!}
                 {!! Form::close() !!}
+            @endcan
         </td>
     </tr>
 @endforeach
 </table>
 {!! $data->render() !!}
-<p class="text-center text-primary"><small>Tutorial by LaravelTuts.com</small></p>
 @endsection
